init 4 files related to web automation.
updatechromedriver will update chromedriver automatically, restart powershell if necessary to take effect, might not need it.

start-autoweb_firefox is specially made for firefox to do incognito windows

start-autoweb_generic is supposed to encompass all of the browers, and however unfinished, chrome with profile works and firefox private and kiosk works, and i think with profile too.